nba-service/seed_requirements.txt
nba_api>=1.4
psycopg2-binary>=2.9.0
numpy>=1.24.0,<2.0
requests>=2.31.0
basketball_reference_web_scraper>=4.9.0
